events_search/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
xlrd==1.2.0
